Fix nuances of DS4
You have to send a "SET_FEATURE" report with 0x05 to switch the
controller to "full" reporting mode. Then, everything arrives in
one huge custom report.

You also need to have SDP enabled (even if there's nothing to discover)
or the controller immediately disconnects.

BUT, btstack needs to be changed to increase the btstack hid_host
MAX_ATTRIBUTE_VALUE_SIZE (512 is sufficient), I think because the report
descriptor is so huge (430 bytes ish)